Multi-service deployment

+++ scripts/host/m_reinstall.sh
@@ -4,12 +4,17 @@ cd "$(dirname "${BASH_SOURCE[0]}")/../.." && vagrant_dir=$PWD
 
 source "${vagrant_dir}/scripts/output_functions.sh"
 
-magento_app_code_dir="${vagrant_dir}/magento2ce/app/code/Magento"
+magento_context="$(bash "${vagrant_dir}/scripts/host/get_magento_context.sh")"
 
-cd "${magento_app_code_dir}"
+instance_path="$(bash "${vagrant_dir}/scripts/get_config_value.sh" "magento_context_${magento_context}_path")"
+magento_ce_dir="${vagrant_dir}/magento/instances/${instance_path}"
 
-status "Deleting TestModule directories"
-ls | grep "TestModule" | xargs rm -rf
+magento_app_code_dir="${magento_ce_dir}/app/code/Magento"
+if [[ -d "${magento_app_code_dir}" ]]; then
+    cd "${magento_app_code_dir}"
+    status "Deleting TestModule directories"
+    ls | grep "TestModule" | xargs rm -rf
+fi
 
 cd "${vagrant_dir}"
 vagrant ssh -c "bash /vagrant/scripts/guest/m-reinstall" 2> >(logError)
